+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Thanks to: @dathbe.
 - [calendar] Fixed broken unittest that only broke on the 1st of July and 1st of january (#3830)
 - [clock] Fixed missing icons when no other modules with icons is loaded (#3834)
 - [weather] Fixed handling of empty values in weathergov providers handling of precipitationAmount (#3859)
+- [calendar] Fixed regression of calendarfetcherutils.shouldEventBeExcluded (#3841)
 
 ## [2.32.0] - 2025-07-01